+++ api/dinosaur/info/peer.java
@@ -4,18 +4,19 @@ public class peer {
 	public String 		ip;
 	public long 		downloaded;
 	public long 		uploaded;
-	public double 		downSpeed;
-	public double 		upSpeed;
+	public long	 		downSpeed;
+	public long	 		upSpeed;
 	public double 		available;
-	public int			blocks2request;
-	public int			requested_blocks;
+	public long			blocks2request;
+	public long			requested_blocks;
 	public boolean 		may_request;
 	public peer()
 	{
 		
 	}
-	public peer(String ip, long downloaded, long uploaded, double downSpeed, double upSpeed, double available,
-				int block2request, int requested_blocks, boolean may_request)
+	public peer(String ip, long downloaded, long uploaded, long downSpeed, 
+			long upSpeed, double available,long block2request, long requested_blocks, 
+			    boolean may_request)
 	{
 		this.ip = ip;
 		this.downloaded = downloaded;
